Rename processing_pipeline.py to pipeline_orchestrator.py for better clarity
