+++ ezstitcher/core/pipeline_orchestrator.py
